fixing a few typing and initialising errors and updating libft.h

+++ ft_strsub.c
@@ -6,21 +6,23 @@
 /*   By: anorman <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/26 13:21:09 by anorman           #+#    #+#             */
-/*   Updated: 2019/05/26 13:28:45 by anorman          ###   ########.fr       */
+/*   Updated: 2019/05/26 16:55:31 by anorman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 char	*ft_strsub(char const *s, unsigned int start, size_t len)
 {
 	char	*substr;
-	int		cnt;
+	size_t	cnt;
 
 	if (!(substr = (char *)(malloc(len * (sizeof(char) + 1)))))
 		return (NULL);
 	cnt = 0;
 	while (cnt < len)
 	{
-		substr[cnt] = s[cnt + start];
+		substr[cnt] = s[cnt + (size_t)start];
 		cnt++;
 	}
 	substr[cnt] = '\0';
